requirements.txt
flask
flask-cors
requests
pillow
numpy
opencv-python
gradio_client
